QWSI: Use floating point number for tablet tilt
On some platforms (like Wayland and X11) we get non-integer values,
and the QTabletEvent API also exposes it as qreal.

However the QWSI API uses int, resulting in potential loss of
information

Task-number: QTBUG-8059

+++ src/gui/kernel/qwindowsysteminterface.cpp
@@ -889,7 +889,7 @@ void QWindowSystemInterfacePrivate::TabletEvent::setPlatformSynthesizesMouse(boo
 
 bool QWindowSystemInterface::handleTabletEvent(QWindow *window, ulong timestamp, const QPointingDevice *device,
                                                const QPointF &local, const QPointF &global,
-                                               Qt::MouseButtons buttons, qreal pressure, int xTilt, int yTilt,
+                                               Qt::MouseButtons buttons, qreal pressure, qreal xTilt, qreal yTilt,
                                                qreal tangentialPressure, qreal rotation, int z,
                                                Qt::KeyboardModifiers modifiers)
 {
@@ -903,7 +903,7 @@ bool QWindowSystemInterface::handleTabletEvent(QWindow *window, ulong timestamp,
 
 bool QWindowSystemInterface::handleTabletEvent(QWindow *window, const QPointingDevice *device,
                                                const QPointF &local, const QPointF &global,
-                                               Qt::MouseButtons buttons, qreal pressure, int xTilt, int yTilt,
+                                               Qt::MouseButtons buttons, qreal pressure, qreal xTilt, qreal yTilt,
                                                qreal tangentialPressure, qreal rotation, int z,
                                                Qt::KeyboardModifiers modifiers)
 {
@@ -914,7 +914,7 @@ bool QWindowSystemInterface::handleTabletEvent(QWindow *window, const QPointingD
 }
 
 bool QWindowSystemInterface::handleTabletEvent(QWindow *window, ulong timestamp, const QPointF &local, const QPointF &global,
-                                               int device, int pointerType, Qt::MouseButtons buttons, qreal pressure, int xTilt, int yTilt,
+                                               int device, int pointerType, Qt::MouseButtons buttons, qreal pressure, qreal xTilt, qreal yTilt,
                                                qreal tangentialPressure, qreal rotation, int z, qint64 uid,
                                                Qt::KeyboardModifiers modifiers)
 {
@@ -925,7 +925,7 @@ bool QWindowSystemInterface::handleTabletEvent(QWindow *window, ulong timestamp,
 }
 
 bool QWindowSystemInterface::handleTabletEvent(QWindow *window, const QPointF &local, const QPointF &global,
-                                               int device, int pointerType, Qt::MouseButtons buttons, qreal pressure, int xTilt, int yTilt,
+                                               int device, int pointerType, Qt::MouseButtons buttons, qreal pressure, qreal xTilt, qreal yTilt,
                                                qreal tangentialPressure, qreal rotation, int z, qint64 uid,
                                                Qt::KeyboardModifiers modifiers)
 {
@@ -936,7 +936,7 @@ bool QWindowSystemInterface::handleTabletEvent(QWindow *window, const QPointF &l
 
 bool QWindowSystemInterface::handleTabletEnterLeaveProximityEvent(QWindow *window, ulong timestamp, const QPointingDevice *device,
                                                                   bool inProximity, const QPointF &local, const QPointF &global,
-                                                                  Qt::MouseButtons buttons, int xTilt, int yTilt,
+                                                                  Qt::MouseButtons buttons, qreal xTilt, qreal yTilt,
                                                                   qreal tangentialPressure, qreal rotation, int z,
                                                                   Qt::KeyboardModifiers modifiers)
 {
@@ -957,7 +957,7 @@ bool QWindowSystemInterface::handleTabletEnterLeaveProximityEvent(QWindow *windo
 
 bool QWindowSystemInterface::handleTabletEnterLeaveProximityEvent(QWindow *window, const QPointingDevice *device,
                                                                   bool inProximity, const QPointF &local, const QPointF &global,
-                                                                  Qt::MouseButtons buttons, int xTilt, int yTilt,
+                                                                  Qt::MouseButtons buttons, qreal xTilt, qreal yTilt,
                                                                   qreal tangentialPressure, qreal rotation, int z,
                                                                   Qt::KeyboardModifiers modifiers)
 {
